Refactoring into three modules
