--- buildSrc/src/main/kotlin/com/patxi/poetimizely/versioning.kt
@@ -1,14 +1,0 @@
-package com.patxi.poetimizely
-
-import java.io.File
-import java.util.Properties
-
-fun poetimizelyVersion(): String = Properties().apply {
-    load(getVersioningFile().inputStream())
-}.getProperty("version")
-
-fun bumpVersion(version: String): Unit = Properties().apply {
-    put("version", version)
-}.store(getVersioningFile().outputStream(), null)
-
-private fun getVersioningFile(): File = File("version.properties")
