+++ core/.editorconfig
