+++ gradle-plugin/.editorconfig
@@ -1,0 +1,2 @@
+[*.{kt, kts}]
+disabled_rules = import-ordering
